requirements/base.txt
# Bleeding edge Django
django==1.9.9

# Configuration
django-environ==0.4.0
whitenoise==3.2.1

# Forms
django-braces==1.9.0
django-crispy-forms==1.6.0
django-floppyforms==1.7.0

# Models
django-model-utils==2.5.2

# Images
Pillow==3.3.1

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.27.0

# Python-PostgreSQL Database Adapter
psycopg2==2.6.2

# Unicode slugification
awesome-slugify==1.6.5
django-autoslug==1.9.3

# Time zones support
pytz==2016.6.1

# Redis support
django-redis==4.4.4
redis>=2.10.0

# Celery
celery==3.1.23

# Your custom requirements go here

# haystack and elasticsearch
django-haystack==2.5.0
elasticsearch=2.4.0
